Assignment:In-Class-Task 2
I just saw the figure, and did the coding without even seeing the codes which you have mentioned in the word file, However I learn the &nbsp; special character. It did gave me a little idea that how much it can be useful for little tiny details in the website thank you for that.